% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{aal}
\alias{aal}
\title{AAL Cortical Atlas}
\usage{
aal()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas for the original Automated Anatomical Labeling parcellation
(Tzourio-Mazoyer et al. 2002).
}
\examples{
aal()
}
\references{
Tzourio-Mazoyer N, Landeau B, Papathanassiou D, et al. (2002).
Automated Anatomical Labeling of Activations in SPM Using a Macroscopic
Anatomical Parcellation of the MNI MRI Single-Subject Brain.
\emph{NeuroImage}, 15(1):273-289.
\doi{10.1006/nimg.2001.0978}
}
\seealso{
Other ggseg_atlases: 
\code{\link{aal2}()},
\code{\link{aal3_cerebellum}()},
\code{\link{aal3_cortical}()},
\code{\link{aal3_subcortical}()}
}
\concept{ggseg_atlases}
