% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{aal2}
\alias{aal2}
\title{AAL2 Cortical Atlas}
\usage{
aal2()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas for the AAL2 parcellation (Rolls et al. 2015) with 41 cortical
regions per hemisphere. Contains both 2D polygon geometry for
\code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and 3D vertex indices for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\details{
Region names confirmed by Herve Lemaitre (AAL2 co-author), mapping the
original generic annotation labels to anatomical AAL2 names.
}
\examples{
aal2()
}
\references{
Rolls, E. T., Joliot, M., & Tzourio-Mazoyer, N. (2015).
Implementation of a new parcellation of the orbitofrontal cortex in
the automated anatomical labeling atlas. Neuroimage, 122, 1-5.
(\href{https://pubmed.ncbi.nlm.nih.gov/26241684/}{PubMed})
}
\seealso{
Other ggseg_atlases: 
\code{\link{aal}()},
\code{\link{aal3_cerebellum}()},
\code{\link{aal3_cortical}()},
\code{\link{aal3_subcortical}()}
}
\concept{ggseg_atlases}
