% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aal3_cortical}
\alias{aal3_cortical}
\title{AAL3 Cortical Atlas}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\usage{
data(aal3_cortical)
}
\description{
Cortical parcellation from AAL3 (Rolls et al. 2020) with 45 regions per
hemisphere. Contains 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}}.
}
\examples{
data(aal3_cortical)
aal3_cortical
}
\references{
Rolls, E. T., Huang, C. C., Lin, C. P., Feng, J., & Joliot, M.
(2020). Automated anatomical labelling atlas 3. Neuroimage, 206, 116189.
(\href{https://pubmed.ncbi.nlm.nih.gov/31521825/}{PubMed})
}
\seealso{
Other ggseg_atlases: 
\code{\link{aal2}},
\code{\link{aal3_cerebellum}},
\code{\link{aal3_subcortical}}
}
\concept{ggseg_atlases}
\keyword{datasets}
