% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{aal3_subcortical}
\alias{aal3_subcortical}
\title{AAL3 Subcortical (Basal Ganglia) Atlas}
\usage{
aal3_subcortical()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Subcortical parcellation from AAL3 (Rolls et al. 2020) covering basal
ganglia and thalamic nuclei with 22 regions. Contains both 2D polygon
geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and 3D meshes for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
aal3_subcortical()
}
\references{
Rolls, E. T., Huang, C. C., Lin, C. P., Feng, J., & Joliot, M.
(2020). Automated anatomical labelling atlas 3. Neuroimage, 206, 116189.
(\href{https://pubmed.ncbi.nlm.nih.gov/31521825/}{PubMed})
}
\seealso{
Other ggseg_atlases: 
\code{\link{aal}()},
\code{\link{aal2}()},
\code{\link{aal3_cerebellum}()},
\code{\link{aal3_cortical}()}
}
\concept{ggseg_atlases}
