% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegAal-package.R
\docType{package}
\name{ggsegAAL-package}
\alias{ggsegAAL}
\alias{ggsegAAL-package}
\title{ggsegAAL: AAL Atlases for the 'ggseg' Ecosystem}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

AAL (Automated Anatomical Labeling) atlases for the 'ggseg' ecosystem. Provides unified 'ggseg_atlas' objects for AAL2 (cortical, Rolls et al. 2015) and AAL3 (cortical, subcortical, cerebellum, Rolls et al. 2020), with both 2D polygon geometry and 3D vertex/mesh data for use with 'ggseg' and 'ggseg3d'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegAal}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegAal/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Other contributors:
\itemize{
  \item Herve Lemaitre \email{herve.lemaitre@u-bordeaux.fr} (\href{https://orcid.org/0000-0002-5952-076X}{ORCID}) [contributor]
}

}
\keyword{internal}
