% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aal}
\alias{aal}
\title{AAL (Automated Anatomical Labeling) Atlas}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\usage{
data(aal)
}
\description{
Brain atlas for the AAL parcellation with 41 regions per hemisphere.
Contains both 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and
3D vertex indices for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
data(aal)
aal
}
\references{
Rolls, E. T., Joliot, M., & Tzourio-Mazoyer, N. (2015).
Implementation of a new parcellation of the orbitofrontal cortex in
the automated anatomical labeling atlas. Neuroimage, 122, 1-5.
(\href{https://pubmed.ncbi.nlm.nih.gov/26241684/}{PubMed})
}
\concept{ggseg_atlases}
\keyword{datasets}
